+++ src/extended_data/containers/__init__.py
@@ -1,5 +1,6 @@
 """Tier 2 extended container classes."""
 
+from extended_data.containers.data import ExtendedData
 from extended_data.containers.factory import extend_data, to_builtin
 from extended_data.containers.mappings import ExtendedDict
 from extended_data.containers.sequences import ExtendedList, ExtendedSet, ExtendedTuple
@@ -7,6 +8,7 @@ from extended_data.containers.strings import ExtendedString
 
 
 __all__ = [
+    "ExtendedData",
     "ExtendedDict",
     "ExtendedList",
     "ExtendedSet",